refactor: add TaskSection
