+++ apps/animated-landing/src/components/TaskCard.tsx
@@ -22,12 +22,14 @@ export default function TaskCard({
   size,
 }: TaskCardProps) {
   return (
-    <article className="text-p3 bg-n1 flex flex-col items-center gap-2 rounded-2xl py-4 px-6">
-      <span className="text-n5 text-ch-title bg-g1 rounded-lg py-1 px-2 font-bold"># {tag}</span>
+    <article className="text-p3 bg-n1 section rounded-2xl py-4 px-6 sm:w-full sm:max-w-[300px] sm:py-5">
+      <span className="text-n5 text-ch-title bg-g1 mb-4 rounded-lg py-1 px-2 font-bold sm:mb-5">
+        # {tag}
+      </span>
       <h3 className="text-ch-h4 font-bold uppercase">{title}</h3>
-      <p className="text-ch-h5 mb-1">{description}</p>
-      <Image src={src} alt={title} width={size} height={size} />
-      <div className="mt-3 flex w-full justify-between">
+      <p className="text-ch-h5 mt-2 flex-1">{description}</p>
+      <Image src={src} alt={title} width={size} height={size} className="my-4" />
+      <div className="flex w-full flex-1 items-end justify-between">
         <p className="font-en text-en-subtitle text-p2 uppercase tracking-widest">{footer}</p>
         <ExternalLink
           className="text-ch-subtitle inline-flex items-center gap-2 font-bold"
